backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["spreadsheet_chat"]

[project]
name = "spreadsheet-chat"
version = "0.1.0"
description = "FastAPI backend: spreadsheet Q&A via LLM-generated openpyxl code"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ai-sdk-python>=0.1.1",
    "fastapi>=0.135.1",
    "openpyxl>=3.1.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
]
